python_scratchpad/pyproject.toml
[tool.poetry]
name = "python-scratchpad"
version = "0.1.0"
description = ""
authors = ["Nicholas Ferrara <nferrara100@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
mypy = "^1.6.1"
requests-mock = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
